Moved .gitignore rules to top level.
This moves all ignore rules to the .gitignore files in the root directory. This is so that users can edit the .gitignore file in src without the risk of merge conflicts if we ever need to edit the default rules.

src/.gitignore is now empty but is left in the repository so that the src directory is automatically created on checkout.